Only call Que.migrate! once
