+++ features/fixtures/que/app/enqueue-job.rb
@@ -1,0 +1,8 @@
+require_relative "setup-que"
+
+case ARGV[0]
+when "unhandled"
+  UnhandledJob.enqueue
+when "handled"
+  HandledJob.enqueue
+end
